requirements.txt
backoff==2.2.1
evaluate==0.4.0
fastcluster==1.2.6
flake8-commas==2.1.0
flake8-isort==5.0.0
flake8==5.0.4
huggingface-hub~=0.19.0
ipykernel==6.19.2
ipywidgets==8.0.3
jupyter-client==8.1.0
jupyter-console==6.4.4
jupyter-server==2.11.2
jupyter==1.0.0
jupyterlab-widgets==3.0.4
keras==2.11.0
keybert~=0.8.0
langdetect==1.0.9
matplotlib-inline==0.1.6
matplotlib==3.6.3
mypy-extensions~=1.0.0
mypy~=1.10.0
numpy==1.26.3
opencage==2.2.0
pandas-stubs>=2.1
pandas~=2.1.0
Pillow~=10.3.0
pre-commit==3.0.3
psycopg2-binary==2.9.6
pyarrow==14.0.1
pycodestyle==2.9.1
PyJWT==2.7.0
pylint~=3.2.0
pytest-cov==4.0.0
pytest==7.1.3
python-dateutil~=2.8.0
pytz>=2023.3
qdrant-client~=1.8.0
qtconsole==5.4.0
quick-server~=0.10.0
redipy~=0.7.0
requests==2.31.0
scattermind~=0.5.0
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
sentence-transformers~=2.3.0
spacy==3.5.3
SQLAlchemy==2.0.9
SQLAlchemy[mypy]==2.0.9
thinc==8.1.12
tokenizers<0.19
tqdm==4.65.0
transformers~=4.38.0
translators~=5.9.0
types-Pillow==9.2.1
types-psycopg2==2.9.21.10
types-python-dateutil~=2.8.0
types-pytz>=2023.3.1.1
types-requests==2.31.0.2
types-urllib3==1.26.25.14
widgetsnbextension==4.0.4
